requirements.txt
flask
flask-cors
python-dotenv
pandas
numpy
gunicorn
